+++ src/app/shared/services/db.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import Dexie from 'dexie';
 import { DB_DEFINITION, DB_NAME } from '../constants/db.constants';
+import { NUMBERS } from '../constants/number.constants';
 
 
 @Injectable({
@@ -14,11 +15,12 @@ export class DatabaseService {
   async createDatabase() {
     this.closeDatabase();
     // Declare tables, IDs and indexes
-    this.db.version(2).stores(DB_DEFINITION);
+    this.db.version(NUMBERS.N_2).stores(DB_DEFINITION);
 
     this.db.open().catch((error) => {
       DatabaseService.swConected = false;
       console.error('Failed to open database: ' + error);
+      //TODO Status service and icon in the header
     }).then(() => {
       DatabaseService.swConected = true;
       console.log('Database opened successfully');
@@ -33,8 +35,20 @@ export class DatabaseService {
     return this.db.table(table).put(data);
   }
 
-  async getData(table: string, id: number) {
+  async getData(table: string, id: number | string) {
     return this.db.table(table).get(id);
   }
+
+  async getAllData(table: string) {
+    return this.db.table(table).toArray();
+  }
+
+  async deleteData(table: string, id: number | string) {
+    return this.db.table(table).delete(id);
+  };
+
+  async deleteAllData(table: string) {
+    return this.db.table(table).clear();
+  };
 
 }